Change `SpanHelpers.Fill<T>` count type to signed
Modify variable and parameter names for clarity and conciseness

+++ src/libraries/System.Private.CoreLib/src/System/String.cs
@@ -297,7 +297,7 @@ namespace System
             string result = FastAllocateString(count);
             if (c != '\0')
             {
-                SpanHelpers.Fill(ref result._firstChar, (uint)count, c);
+                SpanHelpers.Fill(ref result._firstChar, count, c);
             }
             return result;
         }
